factor out transport code into transport.rs
