+++ codex-rs/app-server/src/lib.rs
@@ -8,25 +8,24 @@ use codex_core::config::ConfigBuilder;
 use codex_core::config_loader::CloudRequirementsLoader;
 use codex_core::config_loader::ConfigLayerStackOrdering;
 use codex_core::config_loader::LoaderOverrides;
-use futures::SinkExt;
-use futures::StreamExt;
 use std::collections::HashMap;
 use std::io::ErrorKind;
 use std::io::Result as IoResult;
-use std::net::SocketAddr;
 use std::path::PathBuf;
-use std::str::FromStr;
 use std::sync::Arc;
-use std::sync::atomic::AtomicU64;
-use std::sync::atomic::Ordering;
 
-use crate::message_processor::ConnectionSessionState;
 use crate::message_processor::MessageProcessor;
 use crate::message_processor::MessageProcessorArgs;
 use crate::outgoing_message::ConnectionId;
 use crate::outgoing_message::OutgoingEnvelope;
-use crate::outgoing_message::OutgoingMessage;
 use crate::outgoing_message::OutgoingMessageSender;
+use crate::transport::CHANNEL_CAPACITY;
+use crate::transport::ConnectionState;
+use crate::transport::TransportEvent;
+use crate::transport::has_initialized_connections;
+use crate::transport::route_outgoing_envelope;
+use crate::transport::start_stdio_connection;
+use crate::transport::start_websocket_acceptor;
 use codex_app_server_protocol::ConfigLayerSource;
 use codex_app_server_protocol::ConfigWarningNotification;
 use codex_app_server_protocol::JSONRPCMessage;
@@ -37,18 +36,9 @@ use codex_core::check_execpolicy_for_warnings;
 use codex_core::config_loader::ConfigLoadError;
 use codex_core::config_loader::TextRange as CoreTextRange;
 use codex_feedback::CodexFeedback;
-use tokio::io::AsyncBufReadExt;
-use tokio::io::AsyncWriteExt;
-use tokio::io::BufReader;
-use tokio::io::{self};
-use tokio::net::TcpListener;
-use tokio::net::TcpStream;
 use tokio::sync::mpsc;
 use tokio::task::JoinHandle;
-use tokio_tungstenite::accept_async;
-use tokio_tungstenite::tungstenite::Message as WebSocketMessage;
 use toml::Value as TomlValue;
-use tracing::debug;
 use tracing::error;
 use tracing::info;
 use tracing::warn;
@@ -67,89 +57,9 @@ mod fuzzy_file_search;
 mod message_processor;
 mod models;
 mod outgoing_message;
+mod transport;
 
-/// Size of the bounded channels used to communicate between tasks. The value
-/// is a balance between throughput and memory usage – 128 messages should be
-/// plenty for an interactive CLI.
-const CHANNEL_CAPACITY: usize = 128;
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum AppServerTransport {
-    Stdio,
-    WebSocket { bind_address: SocketAddr },
-}
-
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub enum AppServerTransportParseError {
-    UnsupportedListenUrl(String),
-    InvalidWebSocketListenUrl(String),
-}
-
-impl std::fmt::Display for AppServerTransportParseError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            AppServerTransportParseError::UnsupportedListenUrl(listen_url) => write!(
-                f,
-                "unsupported --listen URL `{listen_url}`; expected `stdio://` or `ws://IP:PORT`"
-            ),
-            AppServerTransportParseError::InvalidWebSocketListenUrl(listen_url) => write!(
-                f,
-                "invalid websocket --listen URL `{listen_url}`; expected `ws://IP:PORT`"
-            ),
-        }
-    }
-}
-
-impl std::error::Error for AppServerTransportParseError {}
-
-impl AppServerTransport {
-    pub const DEFAULT_LISTEN_URL: &'static str = "stdio://";
-
-    pub fn from_listen_url(listen_url: &str) -> Result<Self, AppServerTransportParseError> {
-        if listen_url == Self::DEFAULT_LISTEN_URL {
-            return Ok(Self::Stdio);
-        }
-
-        if let Some(socket_addr) = listen_url.strip_prefix("ws://") {
-            let bind_address = socket_addr.parse::<SocketAddr>().map_err(|_| {
-                AppServerTransportParseError::InvalidWebSocketListenUrl(listen_url.to_string())
-            })?;
-            return Ok(Self::WebSocket { bind_address });
-        }
-
-        Err(AppServerTransportParseError::UnsupportedListenUrl(
-            listen_url.to_string(),
-        ))
-    }
-}
-
-impl FromStr for AppServerTransport {
-    type Err = AppServerTransportParseError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Self::from_listen_url(s)
-    }
-}
-
-#[derive(Debug)]
-enum TransportEvent {
-    ConnectionOpened {
-        connection_id: ConnectionId,
-        writer: mpsc::Sender<OutgoingMessage>,
-    },
-    ConnectionClosed {
-        connection_id: ConnectionId,
-    },
-    IncomingMessage {
-        connection_id: ConnectionId,
-        message: JSONRPCMessage,
-    },
-}
-
-struct ConnectionState {
-    writer: mpsc::Sender<OutgoingMessage>,
-    session: ConnectionSessionState,
-}
+pub use crate::transport::AppServerTransport;
 
 fn config_warning_from_error(
     summary: impl Into<String>,
@@ -454,13 +364,7 @@ pub async fn run_main_with_transport(
                         };
                         match event {
                             TransportEvent::ConnectionOpened { connection_id, writer } => {
-                                connections.insert(
-                                    connection_id,
-                                    ConnectionState {
-                                        writer,
-                                        session: ConnectionSessionState::default(),
-                                    },
-                                );
+                                connections.insert(connection_id, ConnectionState::new(writer));
                             }
                             TransportEvent::ConnectionClosed { connection_id } => {
                                 connections.remove(&connection_id);
@@ -541,305 +445,4 @@ pub async fn run_main_with_transport(
     }
 
     Ok(())
-}
-
-async fn start_stdio_connection(
-    transport_event_tx: mpsc::Sender<TransportEvent>,
-    stdio_handles: &mut Vec<JoinHandle<()>>,
-) -> IoResult<()> {
-    let connection_id = ConnectionId(0);
-    let (writer_tx, mut writer_rx) = mpsc::channel::<OutgoingMessage>(CHANNEL_CAPACITY);
-    transport_event_tx
-        .send(TransportEvent::ConnectionOpened {
-            connection_id,
-            writer: writer_tx,
-        })
-        .await
-        .map_err(|_| std::io::Error::new(ErrorKind::BrokenPipe, "processor unavailable"))?;
-
-    let transport_event_tx_for_reader = transport_event_tx.clone();
-    stdio_handles.push(tokio::spawn(async move {
-        let stdin = io::stdin();
-        let reader = BufReader::new(stdin);
-        let mut lines = reader.lines();
-
-        loop {
-            match lines.next_line().await {
-                Ok(Some(line)) => {
-                    if !forward_incoming_message(
-                        &transport_event_tx_for_reader,
-                        connection_id,
-                        &line,
-                    )
-                    .await
-                    {
-                        break;
-                    }
-                }
-                Ok(None) => break,
-                Err(err) => {
-                    error!("Failed reading stdin: {err}");
-                    break;
-                }
-            }
-        }
-
-        let _ = transport_event_tx_for_reader
-            .send(TransportEvent::ConnectionClosed { connection_id })
-            .await;
-        debug!("stdin reader finished (EOF)");
-    }));
-
-    stdio_handles.push(tokio::spawn(async move {
-        let mut stdout = io::stdout();
-        while let Some(outgoing_message) = writer_rx.recv().await {
-            let Some(mut json) = serialize_outgoing_message(outgoing_message) else {
-                continue;
-            };
-            json.push('\n');
-            if let Err(err) = stdout.write_all(json.as_bytes()).await {
-                error!("Failed to write to stdout: {err}");
-                break;
-            }
-        }
-        info!("stdout writer exited (channel closed)");
-    }));
-
-    Ok(())
-}
-
-async fn start_websocket_acceptor(
-    bind_address: SocketAddr,
-    transport_event_tx: mpsc::Sender<TransportEvent>,
-) -> IoResult<JoinHandle<()>> {
-    let listener = TcpListener::bind(bind_address).await?;
-    let local_addr = listener.local_addr()?;
-    info!("app-server websocket listening on ws://{local_addr}");
-
-    let connection_counter = Arc::new(AtomicU64::new(1));
-    Ok(tokio::spawn(async move {
-        loop {
-            match listener.accept().await {
-                Ok((stream, _peer_addr)) => {
-                    let connection_id =
-                        ConnectionId(connection_counter.fetch_add(1, Ordering::Relaxed));
-                    let transport_event_tx_for_connection = transport_event_tx.clone();
-                    tokio::spawn(async move {
-                        run_websocket_connection(
-                            connection_id,
-                            stream,
-                            transport_event_tx_for_connection,
-                        )
-                        .await;
-                    });
-                }
-                Err(err) => {
-                    error!("failed to accept websocket connection: {err}");
-                }
-            }
-        }
-    }))
-}
-
-async fn run_websocket_connection(
-    connection_id: ConnectionId,
-    stream: TcpStream,
-    transport_event_tx: mpsc::Sender<TransportEvent>,
-) {
-    let websocket_stream = match accept_async(stream).await {
-        Ok(stream) => stream,
-        Err(err) => {
-            warn!("failed to complete websocket handshake: {err}");
-            return;
-        }
-    };
-
-    let (writer_tx, mut writer_rx) = mpsc::channel::<OutgoingMessage>(CHANNEL_CAPACITY);
-    if transport_event_tx
-        .send(TransportEvent::ConnectionOpened {
-            connection_id,
-            writer: writer_tx,
-        })
-        .await
-        .is_err()
-    {
-        return;
-    }
-
-    let (mut websocket_writer, mut websocket_reader) = websocket_stream.split();
-    loop {
-        tokio::select! {
-            outgoing_message = writer_rx.recv() => {
-                let Some(outgoing_message) = outgoing_message else {
-                    break;
-                };
-                let Some(json) = serialize_outgoing_message(outgoing_message) else {
-                    continue;
-                };
-                if websocket_writer.send(WebSocketMessage::Text(json.into())).await.is_err() {
-                    break;
-                }
-            }
-            incoming_message = websocket_reader.next() => {
-                match incoming_message {
-                    Some(Ok(WebSocketMessage::Text(text))) => {
-                        if !forward_incoming_message(&transport_event_tx, connection_id, &text).await {
-                            break;
-                        }
-                    }
-                    Some(Ok(WebSocketMessage::Ping(payload))) => {
-                        if websocket_writer.send(WebSocketMessage::Pong(payload)).await.is_err() {
-                            break;
-                        }
-                    }
-                    Some(Ok(WebSocketMessage::Pong(_))) => {}
-                    Some(Ok(WebSocketMessage::Close(_))) | None => break,
-                    Some(Ok(WebSocketMessage::Binary(_))) => {
-                        warn!("dropping unsupported binary websocket message");
-                    }
-                    Some(Ok(WebSocketMessage::Frame(_))) => {}
-                    Some(Err(err)) => {
-                        warn!("websocket receive error: {err}");
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-    let _ = transport_event_tx
-        .send(TransportEvent::ConnectionClosed { connection_id })
-        .await;
-}
-
-async fn forward_incoming_message(
-    transport_event_tx: &mpsc::Sender<TransportEvent>,
-    connection_id: ConnectionId,
-    payload: &str,
-) -> bool {
-    match serde_json::from_str::<JSONRPCMessage>(payload) {
-        Ok(message) => transport_event_tx
-            .send(TransportEvent::IncomingMessage {
-                connection_id,
-                message,
-            })
-            .await
-            .is_ok(),
-        Err(err) => {
-            error!("Failed to deserialize JSONRPCMessage: {err}");
-            true
-        }
-    }
-}
-
-fn serialize_outgoing_message(outgoing_message: OutgoingMessage) -> Option<String> {
-    let value = match serde_json::to_value(outgoing_message) {
-        Ok(value) => value,
-        Err(err) => {
-            error!("Failed to convert OutgoingMessage to JSON value: {err}");
-            return None;
-        }
-    };
-    match serde_json::to_string(&value) {
-        Ok(json) => Some(json),
-        Err(err) => {
-            error!("Failed to serialize JSONRPCMessage: {err}");
-            None
-        }
-    }
-}
-
-async fn route_outgoing_envelope(
-    connections: &mut HashMap<ConnectionId, ConnectionState>,
-    envelope: OutgoingEnvelope,
-) {
-    match envelope {
-        OutgoingEnvelope::ToConnection {
-            connection_id,
-            message,
-        } => {
-            let Some(connection_state) = connections.get(&connection_id) else {
-                warn!(
-                    "dropping message for disconnected connection: {:?}",
-                    connection_id
-                );
-                return;
-            };
-            if connection_state.writer.send(message).await.is_err() {
-                connections.remove(&connection_id);
-            }
-        }
-        OutgoingEnvelope::Broadcast { message } => {
-            let target_connections: Vec<ConnectionId> = connections
-                .iter()
-                .filter_map(|(connection_id, connection_state)| {
-                    if connection_state.session.initialized {
-                        Some(*connection_id)
-                    } else {
-                        None
-                    }
-                })
-                .collect();
-
-            for connection_id in target_connections {
-                let Some(connection_state) = connections.get(&connection_id) else {
-                    continue;
-                };
-                if connection_state.writer.send(message.clone()).await.is_err() {
-                    connections.remove(&connection_id);
-                }
-            }
-        }
-    }
-}
-
-fn has_initialized_connections(connections: &HashMap<ConnectionId, ConnectionState>) -> bool {
-    connections
-        .values()
-        .any(|connection| connection.session.initialized)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-
-    #[test]
-    fn app_server_transport_parses_stdio_listen_url() {
-        let transport = AppServerTransport::from_listen_url(AppServerTransport::DEFAULT_LISTEN_URL)
-            .expect("stdio listen URL should parse");
-        assert_eq!(transport, AppServerTransport::Stdio);
-    }
-
-    #[test]
-    fn app_server_transport_parses_websocket_listen_url() {
-        let transport = AppServerTransport::from_listen_url("ws://127.0.0.1:1234")
-            .expect("websocket listen URL should parse");
-        assert_eq!(
-            transport,
-            AppServerTransport::WebSocket {
-                bind_address: "127.0.0.1:1234".parse().expect("valid socket address"),
-            }
-        );
-    }
-
-    #[test]
-    fn app_server_transport_rejects_invalid_websocket_listen_url() {
-        let err = AppServerTransport::from_listen_url("ws://localhost:1234")
-            .expect_err("hostname bind address should be rejected");
-        assert_eq!(
-            err.to_string(),
-            "invalid websocket --listen URL `ws://localhost:1234`; expected `ws://IP:PORT`"
-        );
-    }
-
-    #[test]
-    fn app_server_transport_rejects_unsupported_listen_url() {
-        let err = AppServerTransport::from_listen_url("http://127.0.0.1:1234")
-            .expect_err("unsupported scheme should fail");
-        assert_eq!(
-            err.to_string(),
-            "unsupported --listen URL `http://127.0.0.1:1234`; expected `stdio://` or `ws://IP:PORT`"
-        );
-    }
 }
